Add prop for enabling dual variant logo (#20)
Before, in our documentation, logo was available to set in two variants
only on documentation page. It was set through `docusaurus.config.js` in
navbar > logo section.

Now, when new documentation arise - React Native Screens - we need to
have a possiblity to set it also on landing site. To solve that we added
`useLandingLogoDualVariant` prop that defines whether this documentation
use two logo variants on landing. The source for dark and light logos
are set in docusaurus.config.js, just as before.

## Test plan:
1. Install this library in one of the docs 
2. Replace Navbar with the one from this patch and set
`useLandingLogoDualVariant` to `true`
3. Add example logo-dark to navbar section in `docusaurus.config.js` 
```
      navbar: {
        logo: {
          alt: ...
          src: ...
          srcDark:"img/logo-dark.svg"
        },
}
```
4. Run `yarn build && yarn serve`
5. Check if logo (not title!) on Landing changes when you change from
light to dark theme